# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Desk Monitor Platform API
  version: 1.0.0
  description: Contract-first API for the administration console, devices and event producers.
servers:
  - url: /api/v1
...
paths:
...
  /deployments:
    get:
      tags: [Deployments]
      operationId: listDeployments
      security: [{userBearer: []}]
      parameters:
        - $ref: '#/components/parameters/ProductFilter'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Deployment page.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Page'
                  - type: object
                    required: [items]
                    properties:
                      items: {type: array, items: {$ref: '#/components/schemas/Deployment'}}
  /markets/quotes:
    get:
      tags: [Markets]
      operationId: listMarketQuotes
      security: [{userBearer: []}]
      parameters:
        - {name: symbol, in: query, schema: {type: string}}
        - {name: from, in: query, schema: {type: string, format: date-time}}
        - {name: to, in: query, schema: {type: string, format: date-time}}
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Quote history.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Page'
                  - type: object
                    required: [items]
                    properties:
                      items: {type: array, items: {$ref: '#/components/schemas/MarketQuote'}}
  /weather:
    get:
      tags: [Weather]
      operationId: getWeather
      security: [{userBearer: []}]
      parameters:
        - {name: location, in: query, schema: {type: string, default: rio-cuarto}}
      responses:
        '200':
          description: Latest weather observation.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Weather'}
        '404': {$ref: '#/components/responses/NotFound'}
...
components:
  securitySchemes:
    userBearer: {type: http, scheme: bearer, bearerFormat: JWT}
    deviceBearer: {type: http, scheme: bearer, bearerFormat: JWT}
    githubActionsBearer: {type: http, scheme: bearer}
    uptimeBearer: {type: http, scheme: bearer}
    refreshCookie: {type: apiKey, in: cookie, name: desk_refresh}
  parameters:
...
    Cursor: {name: cursor, in: query, schema: {type: string}}
    Limit: {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 25}}
    ProductFilter: {name: product_id, in: query, schema: {type: string}}
...
  responses:
...
    NotFound:
      description: Resource not found.
      content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}
...
  schemas:
    Problem:
      type: object
      required: [type, title, status, request_id]
      properties:
        type: {type: string, format: uri-reference}
        title: {type: string}
        status: {type: integer, minimum: 400, maximum: 599}
        detail: {type: string}
        instance: {type: string, format: uri-reference}
        request_id: {type: string}
        errors: {type: object, additionalProperties: {type: array, items: {type: string}}}
...
    Page:
      type: object
      required: [items, has_more]
      properties:
        items: {type: array, items: {}}
        next_cursor: {type: [string, 'null']}
        has_more: {type: boolean}
...
    Deployment:
      type: object
      required: [id, repository, branch, workflow, status, commit_sha, author, started_at]
      properties:
        id: {type: string}
        product_id: {type: [string, 'null']}
        repository: {type: string}
        branch: {type: string}
        workflow: {type: string}
        status: {type: string, enum: [queued, in_progress, success, failure, cancelled]}
        commit_sha: {type: string}
        commit_message: {type: string}
        author: {type: string}
        started_at: {type: string, format: date-time}
        finished_at: {type: [string, 'null'], format: date-time}
        html_url: {type: [string, 'null'], format: uri}
    MarketQuote:
      type: object
      required: [provider, symbol, quote_asset, last, change_percent, observed_at]
      properties:
        provider: {type: string}
        symbol: {type: string}
        quote_asset: {type: string}
        bid: {type: [number, 'null']}
        ask: {type: [number, 'null']}
        last: {type: number}
        change_percent: {type: number}
        observed_at: {type: string, format: date-time}
        stale: {type: boolean, default: false}
    Weather:
      type: object
      required: [location, temperature_c, apparent_temperature_c, weather_code, observed_at]
      properties:
        location: {type: string}
        temperature_c: {type: number}
        apparent_temperature_c: {type: number}
        humidity_percent: {type: number, minimum: 0, maximum: 100}
        weather_code: {type: integer}
        wind_kmh: {type: number, minimum: 0}
        daily_min_c: {type: number}
        daily_max_c: {type: number}
        precipitation_probability_percent: {type: number, minimum: 0, maximum: 100}
        observed_at: {type: string, format: date-time}
        stale: {type: boolean, default: false}
...